requirements.txt
tensorflow
Keras==2.3.1
fpdf==1.7.2
seaborn==0.9.0
statsmodels==0.11.1
numpy==1.16.5
prefixspan==0.5.2
streamlit==0.62.0
pandas==0.25.1
tsfresh==0.15.1
matplotlib==3.1.1
tqdm==4.44.1
SQLAlchemy==1.3.9
matrixprofile_ts==0.0.9
matrixprofile==1.1.2
Pillow==7.2.0
scikit_learn==0.23.1
